Add support Master-Minion path-regex annotation
