+++ internal/configs/version1/nginx.ingress.tmpl
@@ -102,7 +102,7 @@ server {
 	{{- end}}
 
 	{{range $location := $server.Locations}}
-	location {{  makeLocationPath $location.Path $.Ingress.Annotations | printf }} {
+	location {{  makeLocationPath $location $.Ingress.Annotations | printf }} {
 		set $service "{{$location.ServiceName}}";
 		{{with $location.MinionIngress}}
 		# location for minion {{$location.MinionIngress.Namespace}}/{{$location.MinionIngress.Name}}
