+++ internal/configs/version1/template_helper.go
@@ -14,18 +14,42 @@ func trim(s string) string {
 	return strings.TrimSpace(s)
 }
 
-// makeLocationPath takes a string representing a location path
-// and a map representing Ingress annotations.
+// makeLocationPath takes location and Ingress annotations and returns
+// modified location path with added regex modifier or the original path
+// if if no path-regex annotoation is present in ingressAnnotations
+// or in Location's Ingress.
+//
+// Annotation 'path-regex' set on the Minion takes higher priority comparing
+// to the annotation set on the master (in Master-Minion Ingress setup).
+// If no annotation 'path-regex' is set on Minion and only on Ingress (including master),
+// all location paths will be updated using provided regex modifier types.
+func makeLocationPath(loc *Location, ingressAnnotations map[string]string) string {
+	if loc.MinionIngress != nil {
+		// Case when annotation 'path-regex' set on Location's Minion.
+		_, isMinion := loc.MinionIngress.Annotations["nginx.org/mergeable-ingress-type"]
+		regexType, hasRegex := loc.MinionIngress.Annotations["nginx.org/path-regex"]
+
+		if isMinion && hasRegex {
+			return makePathWithRegex(loc.Path, regexType)
+		}
+	}
+
+	// Case when annotation 'path-regex' set on Ingress (including Master).
+	regexType, ok := ingressAnnotations["nginx.org/path-regex"]
+	if !ok {
+		return loc.Path
+	}
+	return makePathWithRegex(loc.Path, regexType)
+}
+
+// makePathWithRegex takes a path representing a location and a regexType
+// (one of `case_sensitive`, `case_insensitive` or `exact`).
 // It returns a location path with added regular expression modifier.
 // See [Location Directive].
 //
 // [Location Directive]: https://nginx.org/en/docs/http/ngx_http_core_module.html#location
-func makeLocationPath(path string, annotations map[string]string) string {
-	p, ok := annotations["nginx.org/path-regex"]
-	if !ok {
-		return path
-	}
-	switch p {
+func makePathWithRegex(path, regexType string) string {
+	switch regexType {
 	case "case_sensitive":
 		return fmt.Sprintf("~ \"^%s\"", path)
 	case "case_insensitive":
